requirements.txt
pandas
numpy
joblib
networkx
pydantic
scikit-learn
matplotlib
plotly
fastparquet
pyarrow
tqdm
python-dotenv
openpyxl
requests
pytest==7.3.1
